Add files via upload
The purpose of this project was to analyze pregnancy rates within the U.S. using data from Kaggle.com. The goal was to understand if there is a correlation between the decline in women's pregnancy rates and external factors not present in the given data. Examples of such factors could include mental health, environmental conditions, and different income levels. However, I was unable to find data on these factors, so I proceeded to work with the available data.

During the analysis, I observed a trend where women with higher education tended to delay motherhood, potentially focusing on their careers. After conducting further analysis and calculating the mean and median values for each year, I noticed that the numbers were relatively similar. This led me to consider the possibility of data corruption or input errors, which could have affected the accuracy of my deep learning process.

I proceeded to encode the data and used TensorFlow with a sequential application model. However, the accuracy of my model was only 1%. I reviewed the optimization steps and they appeared to be fine, so I suspected that the limited amount of data may have hindered the performance of the neural network. Nonetheless, this project provided valuable learning experiences.

Perhaps in the future, we can further analyze and understand the causes behind the decline in pregnancy rates. Possible contributing factors could include societal changes, a sense of uncertainty about the future, the emergence of a stronger independent mindset, or a shifting perception of the importance of starting a family.