Add files via upload
The data that I was given was from Kaggle.com.

Suicide is an unfortunate outcome that some individuals experience, leading them to end their lives. With the data provided, I aimed to identify any patterns or trends in the rising or declining rates of suicides around the world. I began by organizing and cleaning the data. Interestingly, I observed that from 1985 to 1995, suicide rates increased, but after that period, they started to decrease. This led me to question the factors that might have contributed to this trend. Could it be a decrease in technology usage or a more stable economy? There could be various unknown variables that played a role in the decline.

I compared my findings to other sources such as (https://www.cdc.gov/mmwr/volumes/66/wr/mm6630a6.htm) and found that the data aligned with existing research. My goal was to build a neural network model, but due to the lack of comprehensive data on the root causes of suicide, I was not able to achieve significant success. Nonetheless, I am grateful to have gained knowledge and awareness of this serious issue.

Looking beyond the provided data, I investigated the period from 2016 to the present and noticed a sudden increase in suicide rates. Unfortunately, I couldn't find reliable raw data to validate this observation. It leaves me wondering if the change could be attributed to shifting societal standards, mental health issues, social acceptance, or economic downturns in individuals' lives.

In conclusion, if you or someone you know is facing a tough situation, don't hesitate to seek help. It is a sign of strength to ask for support. Let us all stay connected and spread awareness about this important issue. Thank you for reading.